Fix bugs
When Logged Out. Cannot view product page from a user’s garden.
User should be able to click on a product and be brought to the
products page.
Add before_filter and if statement on product controller and product
show page to fix this bug.

Back button on login page does not work.
Change the back button link to gardens index page.

Both user name and email should be required for registering a new
user.
Add validations to ensure fields are not empty.